+++ dart/constraint/ContactConstraint.cpp
@@ -321,13 +321,6 @@ void ContactConstraint::setConstraintForceMixing(double cfm)
            << "It is set to 1e-9." << std::endl;
     mConstraintForceMixing = 1e-9;
   }
-  if (cfm > 1.0)
-  {
-    dtwarn << "Constraint force mixing parameter[" << cfm
-           << "] is greater than 1.0. "
-           << "It is set to 1.0." << std::endl;
-    mConstraintForceMixing = 1.0;
-  }
 
   mConstraintForceMixing = cfm;
 }