+++ CHANGELOG.md
@@ -33,6 +33,12 @@
   * Replaced M_PI with dart::math::constantsd::pi(): [#1367](https://github.com/dartsim/dart/pull/1367)
   * Enabled octomap support on macOS: [#1078](https://github.com/dartsim/dart/pull/1078)
 
+### [DART 6.9.2 (2019-08-16)](https://github.com/dartsim/dart/milestone/60?closed=1)
+
+* Dynamics
+
+  * Allowed constraint force mixing > 1: [#1371](https://github.com/dartsim/dart/pull/1371)
+
 ### [DART 6.9.1 (2019-06-06)](https://github.com/dartsim/dart/milestone/59?closed=1)
 
 * Collision